+++ modules/swift-commons/SwiftCommonsTests/Sources/Helper/ExpectableTests.swift
@@ -1,6 +1,6 @@
 // ----------------------------------------------------------------------------
 //
-//  ExpectableTests.swift
+//  ExpectTests.swift
 //
 //  @author     Alexander Bragin <alexander.bragin@gmail.com>
 //  @copyright  Copyright (c) 2015, MediariuM Ltd. All rights reserved.
